refactor: settings.ts を lib/firestore/assignment に移設し syncStatus 通知を追加 (#544)
購読4箇所(テーブル/シャッフル/管理者/ペア除外)を createSyncedSubscription でラップして
エラー時に reportSyncError を発火させる。保存・削除は runWriteWithSync でラップして
reportSaveError につなぐ。旧 app/assignment/lib/firebase/settings.ts を削除し、
バレル(index.ts)は @/lib/firestore/assignment から再エクスポートする形に切り替え。
